Modifications to improve data loading speed on very large datasets. We now hold the raw data in arrays of bytes rather than shorts (hopefully reducing memory by 50%). The state table entries store both the raw string (eg A/T) and an array of the values. The string is used for comparisons while parsing the data which is significantly faster than array comparisons.

+++ trunk/src/flapjack/data/GenotypeData.java
@@ -7,7 +7,7 @@ public class GenotypeData
 	// A reference to the chromsome this data applies to
 	private ChromosomeMap map;
 
-	private short[] loci;
+	private byte[] loci;
 
 	public GenotypeData(ChromosomeMap map)
 	{
@@ -17,15 +17,15 @@ public class GenotypeData
 		int size = map.countLoci();
 
 		// Initialize the array to the correct size
-		loci = new short[size];
+		loci = new byte[size];
 	}
 
-	void setLoci(int index, short stateCode)
+	void setLoci(int index, int stateCode)
 	{
-		loci[index] = stateCode;
+		loci[index] = (byte) stateCode;
 	}
 
-	short getState(int index)
+	int getState(int index)
 	{
 		return loci[index];
 	}
@@ -38,6 +38,6 @@ public class GenotypeData
 	/**
 	 * Returns the raw data array for quick access elsewhere (eg rendering).
 	 */
-	public short[] getLociData()
+	public byte[] getLociData()
 		{ return loci; }
 }
